refactor: legacy ui mixins folder structure
